+++ app/assets/javascripts/components/projects/ProjectBoard.js
@@ -1,11 +1,12 @@
 import React, { useEffect } from "react";
 import { connect } from "react-redux";
-import { fetchProjectBoard } from "actions/projectBoard";
+import { fetchProjectBoard, toggleColumn, reverseColumns } from "actions/projectBoard";
 import { fetchPastStories } from "actions/pastIterations";
 import Column from "../Columns/ColumnItem";
 import History from "../stories/History";
 import { getColumns } from "../../selectors/columns";
 import { CHILLY_BIN, DONE, BACKLOG } from '../../models/beta/column';
+import { canCloseColumn } from '../../models/beta/projectBoard';
 import { createStory, closeHistory } from '../../actions/story';
 import AddStoryButton from '../story/AddStoryButton';
 import { status, historyStatus } from 'libs/beta/constants';
@@ -18,7 +19,7 @@ import StorySearch from '../search/StorySearch';
 import SearchResults from './../search/SearchResults';
 import ProjectLoading from './ProjectLoading';
 import SideBar from './SideBar';
-import Columns from '../Columns/Columns';
+import Columns from '../Columns';
 import Stories from "../stories/Stories";
 import Sprints from "../stories/Sprints";
 
@@ -34,7 +35,9 @@ export const ProjectBoard = ({
   chillyBinStories,
   backlogSprints,
   doneSprints,
-  fetchPastStories
+  fetchPastStories,
+  toggleColumn,
+  reverseColumns
 }) => {
   useEffect(() => {
     fetchProjectBoard(projectId)
@@ -53,7 +56,9 @@ export const ProjectBoard = ({
             state: status.UNSCHEDULED
           })}
         />,
-      children: <Stories stories={chillyBinStories} />
+      children: <Stories stories={chillyBinStories} />,
+      visible: projectBoard.visibleColumns.chillyBin,
+      onClose: () => toggleColumn('chillyBin')
     },
     {
       title: `${I18n.t("projects.show.backlog")} / ${I18n.t("projects.show.in_progress")}`,
@@ -63,11 +68,46 @@ export const ProjectBoard = ({
             state: status.UNSTARTED
           })}
         />,
-      children: <Sprints sprints={backlogSprints} />
+      children: <Sprints sprints={backlogSprints} />,
+      visible: projectBoard.visibleColumns.backlog,
+      onClose: () => toggleColumn('backlog')
     },
     {
       title: I18n.t("projects.show.done"),
-      children: <Sprints sprints={doneSprints} fetchStories={fetchPastStories} />
+      children: <Sprints sprints={doneSprints} fetchStories={fetchPastStories} />,
+      visible: projectBoard.visibleColumns.done,
+      onClose: () => toggleColumn('done')
+    }
+  ];
+
+  const sideBarButtons = [
+    {
+      description: I18n.t('revert_columns_tooltip'),
+      onClick: reverseColumns,
+      'data-id': 'reverse-button',
+      isVisible: projectBoard.reverse,
+      icon: 'fas fa-columns'
+    },
+    {
+      description: I18n.t('toggle_column', { column: 'chilly bin' }),
+      onClick: () => toggleColumn('chillyBin'),
+      'data-id': 'toggle-chilly-bin',
+      isVisible: projectBoard.visibleColumns.chillyBin,
+      icon: 'fas fa-snowflake'
+    },
+    {
+      description: I18n.t('toggle_column', { column: 'backlog' }),
+      onClick: () => toggleColumn('backlog'),
+      'data-id': 'toggle-backlog',
+      isVisible: projectBoard.visibleColumns.backlog,
+      icon: 'fas fa-th-list'
+    },
+    {
+      description: I18n.t('toggle_column', { column: 'done' }),
+      onClick: () => toggleColumn('done'),
+      'data-id': 'toggle-done',
+      isVisible: projectBoard.visibleColumns.done,
+      icon: 'fas fa-check-circle'
     }
   ];
 
@@ -80,14 +120,15 @@ export const ProjectBoard = ({
     <div className="ProjectBoard">
       <StorySearch projectId={projectId} loading={projectBoard.search.loading} />
 
-      <SideBar />
+      <SideBar data-id="side-bar" buttons={sideBarButtons} />
 
       <Notifications
         notifications={notifications}
         onRemove={removeNotification}
+        data-id="notifications"
       />
 
-      <Columns {...columnProps} />
+      <Columns {...columnProps} canCloseColumn={canCloseColumn(projectBoard)} />
 
       <SearchResults />
 
@@ -116,8 +157,14 @@ ProjectBoard.propTypes = {
   fetchProjectBoard: PropTypes.func.isRequired,
   createStory: PropTypes.func.isRequired,
   closeHistory: PropTypes.func.isRequired,
-  notifications: PropTypes.array.isRequired
-}
+  notifications: PropTypes.array.isRequired,
+  reverseColumns: PropTypes.func.isRequired,
+  projectId: PropTypes.string.isRequired,
+  removeNotification: PropTypes.func.isRequired,
+  history: PropTypes.object.isRequired,
+  fetchPastStories: PropTypes.func.isRequired,
+  toggleColumn: PropTypes.func.isRequired
+};
 
 const mapStateToProps = ({
   projectBoard,
@@ -149,10 +196,12 @@ const mapStateToProps = ({
 
 const mapDispatchToProps = {
   fetchProjectBoard,
+  toggleColumn,
   createStory,
   closeHistory,
   fetchPastStories,
-  removeNotification
+  removeNotification,
+  reverseColumns
 };
 
 export default connect(