+++ app/assets/javascripts/components/projects/SideBar/SideBarButton.js
@@ -1,19 +1,28 @@
 import React, { useState } from 'react';
 import SideBarButtonInfo from './SideBarButtonInfo';
 import PropTypes from 'prop-types';
+import classname from 'classnames';
 
 const SideBarButton = ({
   children,
   description,
-  onClick
+  onClick,
+  isVisible
 }) => {
   const [showInfo, setShowInfo] = useState(false);
+
+  const classes = classname(
+    'SideBar__link',
+    {
+      'SideBar__link--is-visible': isVisible
+    }
+  );
 
   return (
     <li
       onMouseOver={() => setShowInfo(true)}
       onMouseOut={() => setShowInfo(false)}
-      className="SideBar__link"
+      className={classes}
       onClick={onClick}
       data-id="side-bar-button"
     >
@@ -28,7 +37,8 @@ const SideBarButton = ({
 SideBarButton.propTypes = {
   children: PropTypes.node.isRequired,
   description: PropTypes.string.isRequired,
-  onClick: PropTypes.func.isRequired
+  onClick: PropTypes.func.isRequired,
+  isVisible: PropTypes.bool.isRequired
 }
 
 export default SideBarButton;