+++ app/assets/javascripts/components/shapes/column.js
@@ -4,7 +4,8 @@ const columnShape = PropTypes.shape({
   title: PropTypes.string.isRequired,
   renderAction: PropTypes.func,
   children: PropTypes.node.isRequired,
-  onClose: PropTypes.func
+  onClose: PropTypes.func,
+  visible: PropTypes.bool.isRequired
 });
 
 export default columnShape;