+++ spec/javascripts/components/projects/project_board_spec.js
@@ -10,7 +10,12 @@ describe('<ProjectBoard />', () => {
         search: {
           loading: false
         },
-        reverse: false
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
+        }
       },
       doneSprints: [],
       backlogSprints: [],
@@ -22,11 +27,53 @@ describe('<ProjectBoard />', () => {
         status: 'DISABLED'
       },
       onRemove: sinon.stub(),
-      removeNotification: sinon.stub()
+      removeNotification: sinon.stub(),
+      toggleColumn: sinon.stub(),
+      reverseColumns: sinon.stub(),
+      projectId: '1',
+      fetchPastStories: sinon.stub()
     };
 
     return shallow(<ProjectBoard {...defaultProps} {...props } />);
-  }
+  };
+
+  it('renders <SideBar />', () => {
+    const wrapper = render({
+      projectBoard: { 
+        isFetched: true,
+        search: {
+          loading: false
+        },
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
+        },
+        reverse: true
+      }
+    });
+
+    expect(wrapper.find('[data-id="side-bar"]')).toExist();
+  });
+
+  it('render <Notifications />', () => {
+    const wrapper = render({
+      projectBoard: { 
+        isFetched: true,
+        search: {
+          loading: false
+        },
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
+        },
+        reverse: true
+      },
+    });
+
+    expect(wrapper.find('[data-id="notifications"]')).toExist();
+  });
 
   describe('when projectBoard.isFetched is false',  () => {
     it('renders <ProjectLoading />', () => {
@@ -44,7 +91,13 @@ describe('<ProjectBoard />', () => {
           isFetched: true,
           search: {
             loading: false
-          }
+          },
+          visibleColumns: {
+            backlog: true,
+            done: true,
+            chillyBin: true
+          },
+          reverse: true
         },
       });
       const spinnerLoading = wrapper.find('[data-id="project-loading"]');
@@ -60,9 +113,15 @@ describe('<ProjectBoard />', () => {
         search: {
           loading: false
         },
-        reverse: false
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
+        }
       }
-    }
+    };
+
     it('renders normal-column', () => {
       const wrapper = render(aditionalProps);
       const normalColumn = wrapper.find('[data-id="normal-column"]');
@@ -85,7 +144,12 @@ describe('<ProjectBoard />', () => {
         search: {
           loading: false
         },
-        reverse: true
+        reverse: true,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
+        }
       }
     };
 
@@ -115,6 +179,12 @@ describe('<ProjectBoard />', () => {
         isFetched: true,
         search: {
           loading: false
+        },
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
         }
       },
     };
@@ -149,6 +219,12 @@ describe('<ProjectBoard />', () => {
         isFetched: true,
         search: {
           loading: false
+        },
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
         }
       },
     };
@@ -183,6 +259,12 @@ describe('<ProjectBoard />', () => {
         isFetched: true,
         search: {
           loading: false
+        },
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
         }
       },
     };
@@ -217,6 +299,12 @@ describe('<ProjectBoard />', () => {
         isFetched: true,
         search: {
           loading: false
+        },
+        reverse: false,
+        visibleColumns: {
+          backlog: true,
+          done: true,
+          chillyBin: true
         }
       },
     };